Change type for input index

+++ src/Lykke.Service.Iota.Api.AzureRepositories/AddressInput/AddressInputEntity.cs
@@ -18,6 +18,6 @@ namespace Lykke.Service.Iota.Api.AzureRepositories
             get => RowKey;
         }
 
-        public long Index { get; set; }
+        public int Index { get; set; }
     }
 }
